+++ packages/app/src/pages/plot/implicit-surface.tsx
@@ -18,25 +18,38 @@ import {
 
 let t = 0;
 
-const f = (x: number, y: number, z: number) => Math.sin(x)*Math.cos(y) + Math.sin(y)*Math.cos(z) + Math.sin(z)*Math.cos(x);
-
-const EXPR_POSITION = (emit: Emit, x: number, y: number, z: number) => {
-  emit(x, y, z);
+const f = (x: number, y: number, z: number, t: number) => {
+  x = x * 2 + t;
+  y = y * 2;
+  z = z * 2;
+  return Math.sin(x)*Math.cos(y) + Math.sin(y)*Math.cos(z) + Math.sin(z)*Math.cos(x);
 }
 
-const EXPR_VALUE = (emit: Emit, x: number, y: number, z: number) => {
-  emit(f(x, y, z));
+const EXPR_POSITION = (emit: Emit, x: number, y: number, z: number, i: number, j: number, k: number, time: Time) => {
+  const t = time.elapsed / 1000;
+  emit(x, y, z, t);
 }
 
-const EXPR_NORMAL = (emit: Emit, x: number, y: number, z: number) => {
+const EXPR_VALUE = (emit: Emit, x: number, y: number, z: number, i: number, j: number, k: number, time: Time) => {
+  const t = time.elapsed / 1000;
+  emit(f(x, y, z, t));
+}
+
+const EXPR_NORMAL = (emit: Emit, x: number, y: number, z: number, i: number, j: number, k: number, time: Time) => {
+  const t = time.elapsed / 1000;
   const e = 1e-5;
 
-  const v = f(x, y, z);
-  const vx = f(x + e, y, z);
-  const vy = f(x, y + e, z);
-  const vz = f(x, y, z + e);
+  const v  = f(x, y, z, t);
+  const vx = f(x + e, y, z, t);
+  const vy = f(x, y + e, z, t);
+  const vz = f(x, y, z + e, t);
   
-  emit((vx - x) / e, (vy - y) / e, (vz - z) / e);
+  const nx = (vx - v) / e;
+  const ny = (vy - v) / e;
+  const nz = (vz - v) / e;
+  const nl = 1/Math.sqrt(nx * nx + ny * ny + nz * nz);
+  
+  emit(nx * nl, ny * nl, nz * nl);
 };
 
 export const PlotImplicitSurfacePage: LC = () => {
@@ -93,18 +106,24 @@ export const PlotImplicitSurfacePage: LC = () => {
                 format='vec3<f32>'
                 size={[24, 12, 12]}
                 expr={EXPR_POSITION}
+                time
+                live
                 render={(positions: StorageSource) => (
                   <Sampled
                     axes='xyz'
                     format='f32'
                     size={[24, 12, 12]}
                     expr={EXPR_VALUE}
+                    time
+                    live
                     render={(values: StorageSource) => (
                       <Sampled
                         axes='xyz'
                         format='f32'
                         size={[24, 12, 12]}
                         expr={EXPR_NORMAL}
+                        time
+                        live
                         render={(normals: StorageSource) => [
                           <DualContourLayer
                             values={values}
@@ -112,12 +131,14 @@ export const PlotImplicitSurfacePage: LC = () => {
                             range={[[-3, 3], [-1, 1], [-1, 1]]}
                             color={[0.7, 0.0, 0.5, 1.0]}
                           />,
+                          /*
                           <PointLayer
                             positions={positions}
                             colors={values}
                             size={3}
                             depth={1}
                           />,
+                          */
                         ]}
                       />  
                     )}