+++ packages/shader/src/util/bundle.ts
@@ -44,59 +44,102 @@ export const parseString = (s: string) => s.slice(1, -1).replace(/\\(.)/g, '$1')
 type ToTypeString = (t: TypeLike | string) => string;
 type ToArgTypes = (t: (TypeLike | string)[]) => string[];
 
+export const makeDeclarationToAttribute = (
+  toTypeString: ToTypeString,
+  toArgTypes: ToArgTypes,
+) => (
+  d: any,
+) => {
+  if (d.func) {
+    const {type, name, parameters} = d.func;
+    return {name, format: toTypeString(type), args: toArgTypes(parameters)};
+  }
+  if (d.variable) {
+    const {type, name, parameters} = d.variable;
+    return {name, format: toTypeString(type), args: null};
+  }
+  if (d.struct) {
+    const {name, members} = d.struct;
+    const ms = members?.map(({name, type}: any) => ({
+      name,
+      format: toTypeString(type),
+    }));
+    const args = ms?.map(({format}: any) => format);
+    return {name, format: name, args, members: ms};
+  }
+}
+
+// Replace custom types with a reference to the bundle
+const resolveBundleType = (bundle: ShaderModule, attribute: UniformAttribute) => {
+  const {format} = attribute;
+  if (!format.match(/[A-Z]/)) return attribute;
+  
+  const {libs, module} = bundle as any as ParsedBundle;
+  if (!libs || !module) return attribute;
+
+  const {table: {modules}} = module;
+  if (modules) for (const {name: lib, imports} of modules) {
+    for (const {name, imported} of imports) {
+      if (name === format) {
+        const m = libs[lib];
+        if (m) return {...attribute, format: {...m, entry: imported}};
+      }
+    }
+  }
+
+  return attribute;
+}
+
 // Convert bundle to attributes for its external declarations
 export const makeBundleToAttributes = (
   toTypeString: ToTypeString,
   toArgTypes: ToArgTypes,
-) => (
-  bundle: ShaderModule,
-): UniformAttribute[] => {
-  const module = toModule(bundle);
-  const {table: {declarations}} = module;
+) => {
+  const toAttribute = makeDeclarationToAttribute(toTypeString, toArgTypes);
 
-  const out: UniformAttribute[] = [];
-  for (const {func, flags, variable} of declarations) if (func ?? variable) {
-    if (flags & RF.External) {
-      const {type, name, parameters} = func ?? variable;
-      out.push({name, format: toTypeString(type), args: func ? toArgTypes(parameters) : null});
+  return (
+    bundle: ShaderModule,
+  ): UniformAttribute[] => {
+    const module = toModule(bundle);
+    const {table: {declarations}} = module;
+
+    const out: UniformAttribute[] = [];
+    for (const d of declarations) if (d.func ?? d.variable) {
+      if (d.flags & RF.External) {
+        out.push(resolveBundleType(bundle, toAttribute(d)));
+      }
     }
-  }
 
-  return out;
-}
+    return out;
+  };
+};
 
 // Convert bundle to attribute for entry point (or named declaration)
 export const makeBundleToAttribute = (
   toTypeString: ToTypeString,
   toArgTypes: ToArgTypes,
-) => (
-  bundle: ShaderModule,
-  name?: string,
-): UniformAttribute => {
-  const module = toModule(bundle);
-  const {table: {declarations}} = module;
+) => {
+  const toAttribute = makeDeclarationToAttribute(toTypeString, toArgTypes);
 
-  const entry = name ?? bundle.entry ?? module.entry;
+  return (
+    bundle: ShaderModule,
+    name?: string,
+  ): UniformAttribute => {
+    const module = toModule(bundle);
+    const {table: {declarations}} = module;
 
-  for (const d of declarations) {
-    if (d.func) {
-      const {type, name, parameters} = d.func;
-      if (name === entry) {
-        return {name, format: toTypeString(type), args: toArgTypes(parameters)};
+    const entry = name ?? bundle.entry ?? module.entry;
+
+    for (const d of declarations) {
+      if (
+        d.func?.name === entry ||
+        d.variable?.name === entry
+      ) return resolveBundleType(bundle, toAttribute(d));
+      if (d.struct?.name === entry) {
+        return toAttribute(d);
       }
     }
-    if (d.struct) {
-      const {name, members} = d.struct;
-      if (name === entry) {
-        const ms = members?.map(({name, type}: any) => ({
-          name,
-          format: toTypeString(type),
-        }));
-        const args = ms?.map(({format}: any) => format);
-        return {name, format: name, args, members: ms};
-      }
-    }
-  }
 
-  return {name: name ?? 'main', format: 'void', args: NO_ARGS};
-}
+    return {name: name ?? 'main', format: 'void', args: NO_ARGS};
+  };
+};