+++ packages/workbench/src/hooks/useDerivedSource.ts
@@ -1,0 +1,16 @@
+import type { StorageSource } from '@use-gpu/core';
+
+import { resolve } from '@use-gpu/core';
+import { useOne } from '@use-gpu/live';
+
+export const useDerivedSource = (
+  source: StorageSource,
+  override: Record<string, any>,
+) => {
+  return useOne(() => new Proxy(source, {
+    get: (target, s) => {
+      if (Object.hasOwn(override, s)) return resolve(override[s]);
+      return (target as any)[s];
+    },
+  }) as StorageSource, source);
+};