+++ packages/workbench/src/hooks/useRawSource.ts
@@ -3,7 +3,7 @@ import type { StorageSource, UniformType, TypedArray } from '@use-gpu/core';
 import { useContext, useOne, useMemo, useVersion, useNoContext, useNoOne, useNoMemo, useNoVersion, incrementVersion } from '@use-gpu/live';
 import { makeDataBuffer, uploadBuffer, UNIFORM_ARRAY_DIMS } from '@use-gpu/core';
 
-import { useDeviceContext } from '../providers/device-provider';
+import { useDeviceContext, useNoDeviceContext } from '../providers/device-provider';
 import { useBufferedSize, useNoBufferedSize } from './useBufferedSize';
 
 const NO_OPTIONS: RawSourceOptions = {};
@@ -63,7 +63,7 @@ export const useRawSource = (
 }
 
 export const useNoRawSource = () => {
-  useNoContext(DeviceContext);
+  useNoDeviceContext();
   useNoBufferedSize();
   useNoOne();
   useNoVersion();